Added the Notification.permission raw value.
Notification.permission is a three state value; "default,granted,denied" where "default" is the only state a user will be shown a popup to accept or block notifications.
Added as a simple string, to handle states that might be added in the future.

Removed the dart:html package to notification_api.dart since it depreciated. Made a hacky "isSupported" since that is no longer a function for package:web/web.dart anymore.

Modified the example to show Notification.permission and a simple request button, so end user can test states.
Also added tapStream and moved the "default" values from actionStream to it. Now links will be opened when tapped.
Removed the dart:html package replaced it with package:web/web.dart

Bumped the version to 0.0.6